Added create delete update operations to all tab 2

+++ complete/src/main/java/ppms/service/ClubService.java
@@ -2,10 +2,15 @@ package ppms.service;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import ppms.domain.Club;
+import ppms.domain.Sport;
+import ppms.dto.NewClubDTO;
+import ppms.dto.UpdateClubDTO;
 import ppms.repository.ClubRepository;
 
 @Service
@@ -13,6 +18,12 @@ public class ClubService {
 
   @Autowired
   ClubRepository clubRepository;
+
+  @Autowired
+  SportService sportService;
+
+  @Autowired
+  MembershipService membershipService;
 
   public Club findOne(long id) {
     return clubRepository.findOne(id);
@@ -22,7 +33,7 @@ public class ClubService {
     return clubRepository.findAll();
   }
 
-  public void save(Club club) {
+  public void save(@Valid Club club) {
     clubRepository.save(club);
   }
 
@@ -30,8 +41,24 @@ public class ClubService {
     return clubRepository.count();
   }
 
+  // TODO: exception if not 0
   public void delete(Long id) {
-    clubRepository.delete(id);
+    if (membershipService.getAllMembersOfClub(id).size() == 0) {
+      clubRepository.delete(id);
+    }
   }
 
+  public void save(NewClubDTO clubDTO) {
+    Sport sport = sportService.findOne(clubDTO.getSportId());
+    Club club = new Club(clubDTO, sport);
+    club.setSport(sport);
+    save(club);
+  }
+
+  public void update(UpdateClubDTO clubDTO) {
+    Club club = clubRepository.findOne(clubDTO.getClubId());
+    club.update(clubDTO);
+    save(club);
+
+  }
 }
