+++ complete/src/main/java/ppms/service/NutritionalSupplementService.java
@@ -2,10 +2,14 @@ package ppms.service;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import ppms.domain.NutritionalSupplement;
+import ppms.dto.NewSupplementDTO;
+import ppms.dto.UpdateSupplementDTO;
 import ppms.repository.NutritionalSupplementRepository;
 
 @Service
@@ -13,6 +17,9 @@ public class NutritionalSupplementService {
 
   @Autowired
   NutritionalSupplementRepository nutritionalSupplementRepository;
+
+  @Autowired
+  UsedSupplementsService usedSupplementsService;
 
   public NutritionalSupplement findOne(long id) {
     return nutritionalSupplementRepository.findOne(id);
@@ -22,7 +29,7 @@ public class NutritionalSupplementService {
     return nutritionalSupplementRepository.findAll();
   }
 
-  public void save(NutritionalSupplement sport) {
+  public void save(@Valid NutritionalSupplement sport) {
     nutritionalSupplementRepository.save(sport);
   }
 
@@ -30,8 +37,24 @@ public class NutritionalSupplementService {
     return nutritionalSupplementRepository.count();
   }
 
+  // TODO: exception when not 0
   public void delete(Long id) {
-    nutritionalSupplementRepository.delete(id);
+    if (usedSupplementsService.getAllPeopleUsingThisSupplement(id).size() == 0) {
+      nutritionalSupplementRepository.delete(id);
+    }
+  }
+
+  public void save(NewSupplementDTO supplementDTO) {
+    save(new NutritionalSupplement(supplementDTO));
+
+  }
+
+  public void update(UpdateSupplementDTO supplementDTO) {
+    NutritionalSupplement nutritionalSupplement = nutritionalSupplementRepository.findOne(supplementDTO
+        .getSupplementId());
+    nutritionalSupplement.update(supplementDTO);
+
+    save(nutritionalSupplement);
   }
 
 }
