+++ complete/src/main/java/ppms/service/SportService.java
@@ -2,10 +2,14 @@ package ppms.service;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import ppms.domain.Sport;
+import ppms.dto.NewSportDTO;
+import ppms.dto.UpdateSportDTO;
 import ppms.repository.SportRepository;
 
 @Service
@@ -13,6 +17,9 @@ public class SportService {
 
   @Autowired
   SportRepository sportRepository;
+
+  @Autowired
+  PersonService personService;
 
   public Sport findOne(long id) {
     return sportRepository.findOne(id);
@@ -22,7 +29,7 @@ public class SportService {
     return sportRepository.findAll();
   }
 
-  public void save(Sport sport) {
+  public void save(@Valid Sport sport) {
     sportRepository.save(sport);
   }
 
@@ -30,8 +37,20 @@ public class SportService {
     return sportRepository.count();
   }
 
+  // TODO: exception if not 0
   public void delete(Long id) {
-    sportRepository.delete(id);
+    if (personService.getAllPeopleDoingSport(id).size() == 0) {
+      sportRepository.delete(id);
+    }
   }
 
+  public void save(NewSportDTO sportDTO) {
+    save(new Sport(sportDTO.getName()));
+  }
+
+  public void update(UpdateSportDTO sportDTO) {
+    Sport sport = sportRepository.findOne(sportDTO.getSportId());
+    sport.setName(sportDTO.getName());
+    save(sport);
+  }
 }
